requirements.txt
# Testing framework
pytest==8.0.0
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-rerunfailures==13.0

# HTTP client
httpx==0.27.0

# Database
PyMySQL==1.1.0

# Data parsing
PyYAML==6.0.1

# HTTP library
requests==2.31.0
